+++ intergrated_gui/gui_state.ini
@@ -2,23 +2,23 @@
 checked = True
 camera_dist = 45
 warn_dist = 40
-warning_enabled = False
+warning_enabled = True
 
 [time]
 checked = True
-time_limit = 3
+time_limit = 10
 break_time = 3
-warning_enabled = False
+warning_enabled = True
 
 [posture]
 checked = True
-warn_angle = loose
+warn_angle = strict
 model_path = default
-warning_enabled = False
+warning_enabled = True
 
 [brightness]
 checked = False
 slider_value = 0
-webcam_enabled = False
-color_system_enabled = False
+webcam_enabled = True
+color_system_enabled = True
