+++ util/video_writer.py
@@ -1,6 +1,6 @@
 import cv2
 
-from lib.image_type import ColorImage
+from util.image_type import ColorImage
 
 
 class VideoWriter:
